fix: token detail UI issues, OK-22432
